Derivation with option
